* CompartmentEntry: Made a new class by extracting   VCompartment.FeatureEntry. * VCompartment (addEntry, addTreeRoot, addTreeItem): Refactored   addFeatureEntry.   (addCompartmentEntries): Refactoed addFeatureEntries. * VRequirement (caseRequirementConstraintMembership): Use addEntry(). * VStructure (appendText): Quote creole text. * Visitor (cQuote): Added.
