+++ org.omg.sysml.plantuml/src/org/omg/sysml/plantuml/VCompartment.java
@@ -48,6 +48,7 @@ import org.omg.sysml.lang.sysml.Feature;
 import org.omg.sysml.lang.sysml.FeatureDirectionKind;
 import org.omg.sysml.lang.sysml.FeatureMembership;
 import org.omg.sysml.lang.sysml.FeatureTyping;
+import org.omg.sysml.lang.sysml.FeatureValue;
 import org.omg.sysml.lang.sysml.FlowConnectionUsage;
 import org.omg.sysml.lang.sysml.ItemFlow;
 import org.omg.sysml.lang.sysml.Membership;
@@ -119,153 +120,9 @@ public class VCompartment extends VStructure {
         return null;
     }
 
-    protected static class FeatureEntry implements Comparable<FeatureEntry> {
-        public final Feature f;
-        public final String alias;
-        public final String prefix;
-        public final boolean isInherited;
+    private List<CompartmentEntry> compartmentEntries = new ArrayList<CompartmentEntry>();
 
-        public String getParameterPrefix() {
-            FeatureDirectionKind fdk = f.getDirection();
-            if (fdk == null) return "";
-            switch (fdk) {
-            case IN:
-                return "<b>in</b> ";
-            case OUT:
-                return "<b>out</b> ";
-            case INOUT:
-                return "<b>inout</b> ";
-            }
-            return null; // MUST NOT happen.
-        }
-
-        public boolean isParameter() {
-            return f.getDirection() != null;
-        }
-
-        private static class ToTitle extends SysMLSwitch<String> {
-            public String caseBindingConnector(BindingConnector bc) {
-                return "bindings";
-            }
-            public String caseFlowConnectionUsage(FlowConnectionUsage fcu) {
-                return "flows";
-            }
-            public String caseSuccessionFlowConnectionUsage(SuccessionFlowConnectionUsage sfcu) {
-                return "succession flows";
-            }
-        }
-        private static ToTitle toTitle = new ToTitle();
-
-        public String getTitle() {
-            if (isParameter()) {
-                return "parameters";
-            }
-            String s = toTitle.doSwitch(f);
-            if (s != null) return s;
-            s = SysML2PlantUMLText.getStereotypeName(f);
-            if (s.endsWith("s")) {
-                return s + "es";
-            } else {
-                return s + "s";
-            }
-        }
-
-        private static int featureParameterCompare(Feature f1, Feature f2) {
-            FeatureDirectionKind fdk1 = f1.getDirection();
-            FeatureDirectionKind fdk2 = f2.getDirection();
-            if (fdk1 == null) {
-                if (fdk2 == null) return 0;
-                return 1;
-            }
-            if (fdk2 == null) return -1;
-            if (fdk1.ordinal() < fdk2.ordinal()) return -1;
-            if (fdk1.ordinal() > fdk2.ordinal()) return 1;
-            return 0;
-        }
-
-        private static int featureMetaclassCompare(Feature f1, Feature f2) {
-            if (f1 instanceof AttributeUsage) {
-                if (!(f2 instanceof AttributeUsage)) {
-                    return -1;
-                }
-            } else if (f2 instanceof AttributeUsage) {
-                return 1;
-            } else if (f1 instanceof Usage) {
-                if (!(f2 instanceof Usage)) {
-                    return -1;
-                }
-            } else if (f2 instanceof Usage) {
-                return 1;
-            }
-            
-            String ec1 = f1.eClass().getName();
-            String ec2 = f2.eClass().getName();
-            return ec1.compareTo(ec2);
-        }
-
-        /*
-        private static int featureTypeCompare(Feature f1, Feature f2) {
-            List<Type> ts1 = f1.getType();
-            List<Type> ts2 = f2.getType();
-        }
-        */
-
-        private static int featureNameCompare(Feature f1, Feature f2) {
-            String name1 = f1.getName();
-            String name2 = f2.getName();
-            if (name1 == null) {
-                if (name2 == null) return 0;
-                return -1;
-            }
-            if (name2 == null) return 1;
-            return name1.compareTo(name2);
-        }
-
-        private static int featureCompare(Feature f1, Feature f2) {
-            int v = featureParameterCompare(f1, f2);
-            if (v != 0) return v;
-            v = featureMetaclassCompare(f1, f2);
-            if (v != 0) return v;
-            // The order or EnumerationUsage needs to be stable.
-            if (f1 instanceof EnumerationUsage) return 0;
-            return featureNameCompare(f1, f2);
-        }
-
-        public int compareTo(FeatureEntry o) {
-            int v = featureCompare(f, o.f);
-            if (v != 0) return v;
-            if (alias == null) {
-                if (o.alias == null) return 0;
-                return -1;
-            }
-            if (o.alias == null) return 1;
-            return alias.compareTo(o.alias);
-        }
-
-        public List<FeatureEntry> children;
-        public void add(FeatureEntry fe) {
-            if (children == null) {
-                children = new ArrayList<FeatureEntry>();
-            }
-            children.add(fe);
-        }
-
-        public FeatureEntry(Feature f, String alias, String prefix, boolean isInherited) {
-            this.f = f;
-            this.alias = alias;
-            this.prefix = prefix;
-            this.isInherited = isInherited;
-            this.children = null;
-        }
-    }
-
-    private List<FeatureEntry> featureEntries = new ArrayList<FeatureEntry>();
-
-    protected FeatureEntry addFeature(Feature f,
-                                      String alias,
-                                      String prefix,
-                                      boolean norec,
-                                      FeatureEntry parent) {
+    private CompartmentEntry add(CompartmentEntry ce, CompartmentEntry parent, boolean norec) {
         Membership ms = getCurrentMembership();
         if (ms instanceof ReturnParameterMembership) {
             Element e = ms.getMemberElement();
@@ -273,52 +130,64 @@ public class VCompartment extends VStructure {
             if (e.getOwnedRelationship().isEmpty()) return null;
         }
 
-        if (!norec && (alias == null) && (prefix == null)) {
-            if (recCurrentMembership(f, false)) return null;
+        if (!norec) {
+            if (recCurrentMembership(ce.f, false)) return null;
         }
-        FeatureEntry fe = new FeatureEntry(f, alias, prefix, isInherited());
+
+        if (ce == null) return null;
         if (parent == null) {
-            featureEntries.add(fe);
+            compartmentEntries.add(ce);
         } else {
-            parent.add(fe);
+            parent.add(ce);
         }
-        return fe;
+        return ce;
     }
 
-    protected FeatureEntry addFeature(Feature f,
-                                      String alias,
-                                      String prefix,
-                                      boolean norec) {
-        return addFeature(f, alias, prefix, norec, null);
+    private CompartmentEntry addAlias(Membership ms) {
+        return add(CompartmentEntry.alias(ms, isInherited()), null, true);
     }
 
-    protected FeatureEntry addFeature(Feature f,
-                                      String alias) {
-        return addFeature(f, alias, null, isCompartmentMost());
+    private CompartmentEntry addTreeItem(Membership ms, CompartmentEntry parent) {
+        return add(CompartmentEntry.alias(ms, isInherited()), parent, true);
     }
 
+    private CompartmentEntry addTreeRoot(Feature f) {
+        return add(CompartmentEntry.feature(f, isInherited()), null, true);
+    }
+
+    protected CompartmentEntry addEntry(Feature f, boolean norec) {
+        return add(CompartmentEntry.feature(f, isInherited()), null, norec);
+    }
+
+    protected CompartmentEntry addEntry(Feature f) {
+        return addEntry(f, isCompartmentMost());
+    }
+
+    protected CompartmentEntry addEntry(OwningMembership om, String prefix) {
+        return add(CompartmentEntry.feature(om, prefix, isInherited()), null, true);
+    }
 
     @Override
     public String caseFeature(Feature f) {
-        addFeature(f, null);
+        addEntry(f);
         return "";
     }
     
     @Override
     public String caseConnector(Connector c) {
-    	addFeature(c, null);
+    	addEntry(c);
     	return "";
     }
 
     @Override
     public String caseConnectionUsage(ConnectionUsage cu) {
-    	addFeature(cu, null);
+    	addEntry(cu);
     	return "";
     }
     
     @Override
     public String caseItemFlow(ItemFlow itf) {
-    	addFeature(itf, null);
+    	addEntry(itf);
     	return "";
     }
 
@@ -328,7 +197,7 @@ public class VCompartment extends VStructure {
     }
 
     private class CompTree {
-        private final FeatureEntry parent;
+        private final CompartmentEntry parent;
 
         public void process(Feature f) {
             for (Membership m: f.getOwnedMembership()) {
@@ -339,7 +208,7 @@ public class VCompartment extends VStructure {
                     // Do not show it
                 } else if (e instanceof Feature) {
                 	Feature f2 = (Feature) e;
-                    FeatureEntry fe = addFeature(f2, m.getMemberName(), null, true, parent);
+                    CompartmentEntry fe = addTreeItem(m, parent);
                     CompTree ct = new CompTree(fe);
                     ct.process(f2);
                 } else {
@@ -348,7 +217,7 @@ public class VCompartment extends VStructure {
             }
         }
 
-        private CompTree(FeatureEntry fe) {
+        private CompTree(CompartmentEntry fe) {
             this.parent = fe;
         }
     }
@@ -357,7 +226,7 @@ public class VCompartment extends VStructure {
         boolean compartmentTree = styleValue("compartmentTree") != null;
 
         if (!(isCompartmentMost() || compartmentTree)) return null;
-        FeatureEntry fe = addFeature(f, null, null, true);
+        CompartmentEntry fe = addTreeRoot(f);
         if (compartmentTree) {
             CompTree ct = new CompTree(fe);
             ct.process(f);
@@ -372,7 +241,7 @@ public class VCompartment extends VStructure {
 
     @Override
     public String caseTransitionUsage(TransitionUsage tu) {
-        addFeature(tu, null, null, true);
+        addEntry(tu, true);
         return "";
     }
 
@@ -416,10 +285,7 @@ public class VCompartment extends VStructure {
 
     @Override
     public String caseMembership(Membership m) {
-        Element e = m.getMemberElement();
-        if (e instanceof Feature) {
-            addFeature((Feature) e, m.getMemberName(), null, true);
-        }
+        if (addAlias(m) == null) return null;
         return "";
     }
 
@@ -429,7 +295,7 @@ public class VCompartment extends VStructure {
             Element e = vm.getMemberElement();
             if (e instanceof EnumerationUsage) {
                 EnumerationUsage eu = (EnumerationUsage) e;
-            	addFeature(eu, null, null, true);
+            	addEntry(eu, true);
                 return "";
             }
         }
@@ -477,12 +343,12 @@ public class VCompartment extends VStructure {
                 // Do not show empty ReferenceUsage.
                 if (isEmptyFeature(u)) return "";
             }
-            addFeature(u, null);
+            addEntry(u);
         }
         return "";
     }
 
-    private String addTitle(String prev, FeatureEntry fe) {
+    private String addTitle(String prev, CompartmentEntry fe) {
         String title = fe.getTitle();
         if (title.equals(prev)) return prev;
         append("-- ");
@@ -506,7 +372,7 @@ public class VCompartment extends VStructure {
             Subsetting ss = (Subsetting) e;
             e = ss.getSubsettedFeature();
         }*/
-        append(getText(e));
+        appendText(getText(e), true);
     }
 
     private void addConnectorText(Connector c, boolean isInherited) {
@@ -579,12 +445,12 @@ public class VCompartment extends VStructure {
         return true;
     }
 
-    private void addFeatures(List<FeatureEntry> es, int level) {
+    private void addCompartmentEntries(List<CompartmentEntry> es, int level) {
         Collections.sort(es);
         final int size = es.size();
         String title = null;
         for (int i = 0; i < size; i++) {
-            FeatureEntry fe = es.get(i);
+            CompartmentEntry fe = es.get(i);
             if (level == 0) {
                 title = addTitle(title, fe);
             } else {
@@ -631,7 +497,7 @@ public class VCompartment extends VStructure {
             } else if (addFeatureText(fe.f, false)) {
                 boolean first = true;
                 for (int j = i + 1; j < size; j++) {
-                    FeatureEntry fe2 = es.get(j);
+                    CompartmentEntry fe2 = es.get(j);
                     if (fe.f.equals(fe2.f)) {
                         if (fe2.alias != null) {
                             if (first) append(" <b>alias</b> ");
@@ -648,7 +514,7 @@ public class VCompartment extends VStructure {
             }
 
             if (fe.children != null) {
-                addFeatures(fe.children, level + 1);
+                addCompartmentEntries(fe.children, level + 1);
             }
         }
     }
@@ -656,14 +522,14 @@ public class VCompartment extends VStructure {
     public void startType(Type typ) {
         this.currentType = typ;
         traverse(typ, false, true);
-        addFeatures(featureEntries, 0);
+        addCompartmentEntries(compartmentEntries, 0);
         popNamespace();
     }
 
     public List<VTree> process(VTree parent, Type typ) {
         this.parent = parent;
         subtrees.clear();
-        featureEntries.clear();
+        compartmentEntries.clear();
         startType(typ);
         return subtrees;
     }
@@ -707,5 +573,11 @@ public class VCompartment extends VStructure {
     public String caseActionUsage(ActionUsage au) {
         if (!mixedMode) return null;
         return recCurrent(au, true);
+    }
+
+    @Override
+    public String caseFeatureValue(FeatureValue fv) {
+        addEntry(fv, null);
+        return "";
     }
 }
