+++ org.omg.sysml.plantuml/src/org/omg/sysml/plantuml/VRequirement.java
@@ -75,8 +75,7 @@ public class VRequirement extends VCompartment {
             default:
                 return "";
             }
-            ConstraintUsage c = rcm.getOwnedConstraint();
-            addFeature(c, null, prefix, true);
+            addEntry(rcm, prefix);
         }
         return "";
     }
